%n added

+++ assembler.h
@@ -62,7 +62,7 @@ void addInfo(uint8_t **code);
  * @param line line of code
  * @param commandSize len of command in symbols
  */
-void skipSpaces(Program *program, size_t line, size_t *commandSize);
+void skipSpaces(Program *program, size_t line, int *commandSize);
 
 /**
  * @brief process code if there is [] in line of code
@@ -76,7 +76,7 @@ void skipSpaces(Program *program, size_t line, size_t *commandSize);
  */
 void detectBrackets(Program *program,
                     uint8_t *code,
-                    size_t commandSize,
+                    int commandSize,
                     char *buffer,
                     size_t line,
                     size_t *error);
